Safari-609.1.11.1/LayoutTests/fast/frames/sandboxed-iframe-scripting-03.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test-pre.js"></script>
    <script>
        var allowedExecuted = 0;
        var disallowedExecuted = 0;
        window.onload = function() {
            shouldBe("disallowedExecuted", "0");
            shouldBe("allowedExecuted", "0");

            isSuccessfullyParsed();
        };
    </script>
</head>
<body>
    <iframe sandbox="allow-same-origin"
            src="javascript: ++window.top.disallowedExecuted;">
    </iframe>
    <script>
        description("Verify that sandboxed frames without sandbox='allow-scripts' cannot execute script.");
    </script>
</body>
</html>
