third_party/blink/web_tests/platform/generic/fast/frames/sandboxed-iframe-scripting-03-expected.txt

CONSOLE ERROR: Blocked script execution in 'about:blank' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
Verify that sandboxed frames without sandbox='allow-scripts' cannot execute script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS disallowedExecuted is 0
PASS allowedExecuted is 0
PASS successfullyParsed is true

TEST COMPLETE
